chore(README.md): update level number and description in the README file
The level number in the README file was updated from Level 2 to Level 3 to reflect the correct level of the Devcode Starter using Svelte project. Additionally, the description of the expected outcome was modified to clarify that participants should be able to create a new contact by sending the data to the API and display it in the contact list after it has been created.

fix(services/index.js): uncomment code to import axios and get API URL

The code to import the axios library and get the API URL for fetching contact data from the API was uncommented. This was necessary to ensure that the getAllContactsData function can make successful API requests.

feat(services/index.js): add function to send new contact data to API

A new function, addNewContact, was added to the services/index.js file. This function is responsible for sending the data of a new contact to the API. It uses the axios library to make a POST request to the appropriate API endpoint. Any errors that occur during the request are caught and returned.

fix(components/ContactItem.svelte): remove unnecessary whitespace

An unnecessary whitespace character was removed from the ContactItem component in the components/ContactItem.svelte file. This was done to improve the readability and consistency of the code.

fix(components/HelloWorld.svelte): fix import statement

The import statement in the HelloWorld component in the components/HelloWorld.svelte file was fixed. The path to the devcode-logo.png asset was updated to use single quotes instead of double quotes.

fix(components/InputContactForm.svelte): uncomment code and add event handlers

Several lines of code in the InputContactForm component in the components/InputContactForm.svelte file were uncommented. These lines of code include the import statement for the addNewContact function from the services/index.js file and the handleGetContacts function from the props. Additionally, event handlers were added to the input fields to update the input object when the values change.

feat(components/InputContactForm.svelte): add form submission functionality

The handleSubmit function in the InputContactForm component in the components/InputContactForm.svelte file was implemented. This function is called when the form is submitted and it sends the data of the new contact to the API using the addNewContact function. After the contact has been successfully added, the handleGetContacts function is called to fetch all contact data from the API and update the contactsList variable.